requirements.txt
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
requests>=2.31.0
msal>=1.26.0
